% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{aflhelp}
\alias{aflhelp}
\title{Display SciDB AFL operator documentation}
\usage{
aflhelp(topic)
}
\arguments{
\item{topic}{a SciDB operator or macro name as character string or \code{\link{afl}} object}
}
\value{
displays help
}
\description{
Display SciDB AFL operator documentation
}
\examples{
\dontrun{
s = scidbconnect()
aflhelp("list")     # explicitly look up a character string
help(s$list)        # same thing via R's \\code{help} function
}
}

