% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{antijoin}
\alias{antijoin}
\title{Antijoin}
\usage{
antijoin(array1, array2)
}
\arguments{
\item{array1}{SciDB array}

\item{array2}{SciDB array}
}
\value{
A single-attribute SciDB array
equal to 1 where the corresponding coordinates
exist in array1, or 2 if those coordinates exist in array2
}
\description{
Given two arrays of same dimensionality, return any coordinates that do NOT
join. For all coordinates, the single attribute shall equal to 1 if those
coordinates exist in array1, or 2 if those coordinates exist in array2.
}

