% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{array_stats}
\alias{array_stats}
\title{SciDB array statistics}
\usage{
array_stats(scidb_array, per_instance = FALSE, per_attribute = FALSE)
}
\arguments{
\item{scidb_array}{either a character name of a stored array or a \code{scidb} array object}

\item{per_instance}{set to \code{TRUE} to break down storage stats by SciDB instance}

\item{per_attribute}{set to \code{TRUE} to break down stats by attribute}
}
\value{
A data frame containinf the statistics
export
}
\description{
Display array storage statistics
}

