% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as.R}
\alias{as.R}
\title{Download SciDB data to R}
\usage{
as.R(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} object (a SciDB array or expression)}
}
\value{
An R \code{\link{data.frame}}
}
\description{
Download SciDB data to R
}
\note{
This convenience function is equivalent to running \code{iquery(db, x, return=TRUE)} for
a SciDB connection object \code{s}.
}
\examples{
\dontrun{
s = scidbconnect()
x = scidb(s, "build(<v:double>[i=1:5], sin(i))")
as.R(x)
#  i          v
#1 1  0.8414710
#2 2  0.9092974
#3 3  0.1411200
#4 4 -0.7568025
#5 5 -0.9589243
}
}

