% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{c,scidb-method}
\alias{c,scidb-method}
\title{Concatenate two SciDB arrays}
\usage{
\S4method{c}{scidb}(x, y)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{y}{\code{scidb} array object}
}
\value{
\code{scidb} array object
}
\description{
Concatenate two SciDB arrays with common schema along their last dimension.
}

