% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{cache}
\alias{cache}
\title{Cache data in a SciDB server-side file}
\usage{
cache(x, name, host = .scidbenv$host, port = .scidbenv$port,
  protocol = .scidbenv$protocol)
}
\arguments{
\item{x}{a raw R binary value}

\item{name}{optional key name}

\item{host}{optional SciDB shim host IP address or name}

\item{port}{optional SciDB shim port number}

\item{protocol}{optional SciDB shim protocol (http or https)}
}
\value{
A character URL that may be used to retrieve the data.
}
\description{
Cache data in a SciDB server-side file
}
\examples{
\dontrun{
scidbconnect()
a <- cache(serialize(head(iris), NULL))
print(a)
unserialize(GET_RAW(a))

# Use a specified key, including a directory name
b <- cache(serialize(head(cars), NULL), "test/cars")
print(b)
unserialize(GET_RAW(b))
}
}
\seealso{
GET_RAW
}

