% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{$.scidb}
\alias{$.scidb}
\title{Projection onto an array attribute (variable)}
\usage{
\method{$}{scidb}(x, ...)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{...}{optional argument \code{drop=FALSE} not yet supported, but will be in a future version}
}
\value{
\code{scidb} array object
}
\description{
Use data frame dollar sign notation to project onto a single array attribute
}
\seealso{
\code{\link{project}}
}

