% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{cbind.scidb}
\alias{cbind.scidb}
\title{Combine attributes from two SciDB arrays}
\usage{
cbind.scidb(x, y)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{y}{\code{scidb} array object}
}
\value{
\code{scidb} array object
}
\description{
Concatenate arrays with common dimension schema by combining their attributes.
Equivalent to SciDB's "merge" operator, or \code{\link{aggregate}} with \code{merge=TRUE}.
}

