% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\name{cor}
\alias{cor}
\title{Correlation matrix
This function is more limited than R's default \code{cor} function. It can
only compute a correlation matrix from a data matrix without any missing value
handling.}
\usage{
cor(x, y = NULL, use = "everything", method = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{x}{a 2-d scidb array with a single numeric attribute}
}
\value{
correlation matrix of \code{x} (as a SciDB array)
}
\description{
Correlation matrix
This function is more limited than R's default \code{cor} function. It can
only compute a correlation matrix from a data matrix without any missing value
handling.
}

