% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{count}
\alias{count}
\title{Count the number of non-empty cells in a SciDB array}
\usage{
count(x)
}
\arguments{
\item{x}{\code{scidb} array object}
}
\value{
count of non-empty cells
}
\description{
Count the number of non-empty cells in a SciDB array
}

