% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\name{cov}
\alias{cov}
\title{Covariance matrix
This function is more limited than R's default \code{cov} function. It can
only compute a covariance matrix from a data matrix without any missing value
handling by the procedure (in R notation)
\code{S0 <- sweep(x, 2, colMeans(x))}, 
\code{crossprod(S0)/(nrow(S0) - 1)   # (covariance matrix result)}}
\usage{
cov(x, y = NULL, use = "everything", method = c("pearson", "kendall",
  "spearman"))
}
\arguments{
\item{x}{a 2-d scidb array with a single numeric attribute}
}
\value{
covariance matrix of \code{x} (as a SciDB array)
}
\description{
Covariance matrix
This function is more limited than R's default \code{cov} function. It can
only compute a covariance matrix from a data matrix without any missing value
handling by the procedure (in R notation)
\code{S0 <- sweep(x, 2, colMeans(x))}, 
\code{crossprod(S0)/(nrow(S0) - 1)   # (covariance matrix result)}
}

