% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{dim<-.scidb}
\alias{dim<-.scidb}
\title{Set SciDB array dimension shape}
\usage{
\method{dim}{scidb}(x) <- value
}
\arguments{
\item{x}{a \code{scidb} object}

\item{value}{a vector of dimension bounds}
}
\value{
a \code{scidb} object
}
\description{
This function uses \code{\link{reshape_scidb}} to change array
dimensionality. Does not work with unbounded arrays.
}
\note{
The product of the vector of new dimension bounds must equal the product of the old dimension bounds.
}
\seealso{
\code{\link{reshape_scidb}}
}

