% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{dimension_rename}
\alias{dimension_rename}
\title{Rename SciDB array dimensions}
\usage{
dimension_rename(x, old, new)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{old}{Either a numeric vector of dimensions positions, or a character vector of dimension names}

\item{new}{Character vector of new dimension names of same length as \code{old}}
}
\value{
a new \code{scidb} array object
}
\description{
Rename SciDB array dimensions
}

