% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{dimensions}
\alias{dimensions}
\title{SciDB array dimension names}
\usage{
dimensions(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character vector of SciDB dimension names
}
\description{
SciDB array dimension names
}

