% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.Matrix2scidb}
\alias{.Matrix2scidb}
\title{Internal function to upload an R sparse matrix into SciDB}
\usage{
.Matrix2scidb(
  db,
  X,
  name,
  rowChunkSize = 1000,
  colChunkSize = 1000,
  start = c(0, 0),
  temp = FALSE,
  gc = TRUE,
  ...
)
}
\arguments{
\item{db}{scidb database connection}

\item{X}{a sparse matrix}

\item{name}{(character) SciDB array name}

\item{start}{(int) dimension start values}

\item{temp}{(boolean) create a temporary SciDB array}

\item{gc}{(boolean) set to \code{TRUE} to connect SciDB array to R's garbage collector}

\item{rowChuckSize, colChunkSize}{(int) optional value passed to the aio_input operator see https://github.com/Paradigm4/accelerated_io_tools}
}
\value{
a \code{\link{scidb}} object
}
\description{
Internal function to upload an R sparse matrix into SciDB
}
\keyword{internal}
