% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{filter_scidb}
\alias{filter_scidb}
\title{Flexible SciDB array subsetting}
\usage{
filter_scidb(x, expr)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{expr}{Either a quoted SciDB filter expression, or an R expression involving array attributes and dimensions}
}
\value{
a new \code{scidb} array object
}
\description{
Flexible SciDB array subsetting
}
\note{
The \code{expr} value can include scalar R values, but not more complicated expressions since the expression
is evaluated on the server inside SciDB (not R). Scalar R values are translated to constants in the SciDB expression
using R dynamic scoping/nonstandard evaluation (NSE). Quote full expressions to avoid NSE and force evaluation of
the quoted expression in SciDB (see examples).
}
\keyword{internal}

