% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gemm}
\alias{gemm}
\title{SciDB dense matrix multiply}
\usage{
gemm(X, Y, Z)
}
\arguments{
\item{X}{a \code{scidb} array object}

\item{Y}{a \code{scidb} array object}

\item{Z}{(optional) \code{scidb} array object}
}
\value{
a \code{scidb} array object corresponding to \code{X \%*\% Y + Z}
}
\description{
SciDB dense matrix multiply
}

