% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gesvd}
\alias{gesvd}
\title{SciDB dense matrix SVD}
\usage{
gesvd(X, type = c("left", "values", "right"))
}
\arguments{
\item{X}{a \code{scidb} array object}

\item{type}{a character vector either "left", "values" or "right" to return the
left singular vectors, singular values, or right singular vectors, respectively.}
}
\value{
a \code{scidb} reference object containing the computed result
}
\description{
SciDB dense matrix SVD
}

