% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{\%as\%,scidb-method}
\alias{\%as\%,scidb-method}
\title{AFL arrary aliasing}
\usage{
\S4method{\%as\%}{scidb}(x, y)
}
\arguments{
\item{x}{an object of class \code{\link{scidb}} (a scidb array or expression)}

\item{y}{alias name}
}
\value{
a \code{\link{scidb}} object
}
\examples{
\dontrun{
s = scidbconnect()
x = scidb(s, "build(<v:double>[i=1:2,1,0], i)"
x \%as\% y
}
}

