% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grouped_aggregate}
\alias{grouped_aggregate}
\title{SciDB grouped aggregate operator}
\usage{
grouped_aggregate(scidb_array, by, FUN)
}
\arguments{
\item{scidb_array}{either a character name of a stored array or a \code{scidb} array object}

\item{by}{a list or character vector of array dimension and/or attribute names to group by}

\item{FUN}{a valid SciDB aggregation function (expressed as a character string)}
}
\value{
A \code{scidb} array reference object
}
\description{
Apply a function to a SciDB array grouped by an array attribute or dimension.
}

