% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{head,scidb-method}
\alias{head,scidb-method}
\title{Return the first part of a SciDB array}
\usage{
\S4method{head}{scidb}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{scidb} object}
}
\value{
a data frame with the first part of the array data
}
\description{
Return the first part of a SciDB array
}

