% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{hist,scidb-method}
\alias{hist,scidb-method}
\title{Histogram}
\usage{
\S4method{hist}{scidb}(x, breaks = 10, right = FALSE, materialize = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{scidb} object with a single numeric SciDB attribute}

\item{breaks}{a single number giving the number of cells for the histogram}

\item{right}{logical; if \code{TRUE}, the histogram cells are right-closed (left open) intervals}

\item{materialize}{logical; if \code{TRUE}, return the histogram object to R}

\item{plot}{logical; if \code{TRUE} plot the output}

\item{...}{further arguments and graphical parameters passed to plotting routinesif \code{plot=TRUE}}
}
\value{
an R histogram object
}
\description{
Return and optionall plot a histogram from SciDB array values
}
\seealso{
\code{\link{hist}}
}

