% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{methods}
\name{image,scidb-method}
\alias{image,scidb-method}
\title{Display a SciDB array as an image}
\usage{
\S4method{image}{scidb}(x, grid = c(500, 500), op = "sum", na = 0, ...)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{grid}{desired image resolution}

\item{op}{downsampling aggregation operator}

\item{na}{replace missing values with this value}

\item{...}{additional arguments for the generic \code{image} method}
}
\value{
An R matrix with the image values and the side-effect of displaying the image
plot of the matrix, unless \code{plot=FALSE} is specified in the optional arguments.
}
\description{
Display a 2-d SciDB matrix as an image, downsampling the image in the database to
the specified resolution.
}

