% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{iqdf}
\alias{iqdf}
\title{Head-like SciDB array inspection}
\usage{
iqdf(x, n = 6L, prob = 1)
}
\arguments{
\item{x}{SciDB array}

\item{n}{maximum number of rows to return}

\item{prob}{sampling probability}
}
\description{
Return the first part of an unpacked SciDB array as a data frame.
}
\note{
Setting the sampling probability low for huge arrays will improve performance.
}
\seealso{
\code{\link{head}}
}

