% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{is.scidb,scidb-method}
\alias{is.scidb,scidb-method}
\title{Test if an object has class "scidb"}
\usage{
\S4method{is.scidb}{scidb}(x)
}
\arguments{
\item{x}{a \code{scidb} object}
}
\value{
\code{TRUE} if \code{x} has class "scidb"
}
\description{
Test if an object has class "scidb"
}

