% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.temp}
\alias{is.temp}
\title{Is the object a temporary SciDB array}
\usage{
is.temp(name)
}
\arguments{
\item{name}{a character name of a SciDB array or a \code{scidb} object}
}
\value{
logical value indicating if \code{name} is temporary
}
\description{
Is the object a temporary SciDB array
}

