% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{logical_plan}
\alias{logical_plan}
\title{SciDB logical plan}
\usage{
logical_plan(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
a character value containing the SciDB logical plan for the object
}
\description{
SciDB logical plan
}

