% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{matvec2scidb}
\alias{matvec2scidb}
\title{Internal function to upload an R vector, dense n-d array or matrix to SciDB}
\usage{
matvec2scidb(db, X, name = tmpnam(db), start, gc = TRUE, temp = FALSE, ...)
}
\arguments{
\item{db}{scidb database connection}

\item{X}{a vector, dense n-d array or matrix}

\item{name}{(character) SciDB array name}

\item{start}{(int) dimension start value}

\item{gc}{(boolean) set to TRUE to connect SciDB array to R's garbage collector}

\item{temp}{(boolean) create a temporary SciDB array}

\item{...}{optional extra arguments.
\itemize{
\item {attr: } {attribute name}
\item {reshape: } {(boolean) to control reshape}
\item {type: } {(character) desired data type - however, limited type conversion available}
\item {max_byte_size: } {(int) maximum size of each block (in bytes) while uploading vectors in
a multi-part fashion. Minimum block size is 8 bytes. }
}}
}
\value{
a \code{\link{scidb}} object
}
\description{
Internal function to upload an R vector, dense n-d array or matrix to SciDB
}
\keyword{internal}
