% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{merge.scidb}
\alias{merge.scidb}
\title{SciDB \code{merge}, \code{cross_join}, and \code{join} operations.}
\usage{
\method{merge}{scidb}(x, y, by = intersect(dimensions(x), dimensions(y)), ...)
}
\arguments{
\item{x}{A \code{scidb} array}

\item{y}{A \code{scidb} array}

\item{by}{optional vector of common dimension or attribute names or dimension indices
to join on. See details below.}

\item{...}{optional additional agruments:
\code{fillin} is an optional argument specifying a value used to fill
           attributes as required by merge, it defaults to null;
\code{merge} if \code{TRUE}, perform a SciDB merge operation instead of join;
\code{all.x} optional argument that, if TRUE, indicates left-outer join;
\code{all.y} optional argument that, if TRUE, indicates right-outer join;
\code{all} is an optional argument that, if TRUE, indicates outer join. It only
           applies in limited settings (equivalent to \code{all.x = TRUE, all.y = TRUE});
\code{by.x} optional vector of dimension or attribute names or dimension indices
               of array \code{x} to join on;
\code{by.y} optional vector of dimension or attribute names or dimension indices
                  of array \code{y} to join on.}
}
\value{
a \code{scidb} object
}
\description{
Only one of either \code{by} or both \code{by.x} and \code{by.y} may be
specified.  If none of the \code{by.x},\code{by.y} arguments are specified, and
\code{by=NULL} the result is the Cartesian cross product of \code{x} and
\code{y}.
}
\details{
The default value of \code{by=intersect(dimensions(x), dimensions(y)}
performs a \code{cross_join}, \code{equi_join},
or \code{join} along common array dimensions only. If you want to perform
a join across all columns (dimensions and SciDB array attributes), then
use \code{by=intersect(names(x), names(y))}.

If only \code{by} is specified, the dimension names or attribute names in
\code{by} are assumed to be common across \code{x} and \code{y}.  Otherwise
dimension names and attribute names are matched across the names listed in
\code{by.x} and \code{by.y}, respectively.

If dimension names are specified and \code{by} contains all the dimensions
in each array, then the SciDB \code{join} or \code{merge} operator is used, otherwise SciDB's
\code{cross_join} or \code{equi_join} operator is used. In each either case, the output is a cross
product set of the two arrays along the specified dimensions.

If \code{by} or any of \code{by.x} and \code{by.y} contain an attribute name
name, the optional SciDB plugin operator \code{equi_join} will be used, see
https://github.com/Paradigm4/equi_join.

Specify \code{merge=TRUE} to perform a SciDB merge operation instead
of a SciDB join.

If \code{all=FALSE} (the default), then a SQL-like `natural join`
(an inner join) is performed. If \code{all=TRUE} then SQL-like `outer join`
is performed, but this case has some limitiations; in particular the
outer join is not available yet for the \code{merge=TRUE} case yet. Specifiy
\code{all.x=TRUE} or \code{all.y=TRUE} to perform left- or right-outer
joins, respectively.

Use the optional \code{equi_join=TRUE} argument to force use of the
equi_join SciDB plugin operator, or set \code{equi_join=FALSE} to force
the use of cross_join. The default value is TRUE.

The various SciDB \code{join} operators generally require that the arrays have
identical partitioning (coordinate system bounds, chunk size, etc.) in the
common dimensions.  The \code{merge} method attempts to rectify SciDB
arrays along the specified dimensions as required before joining. Those
dimensions must at least have common lower index bounds.

The merge function may rename SciDB attributes and dimensions as required
to avoid name conflicts in SciDB. See the last example for an
illustration.
}
\examples{
\dontrun{
x <- as.scidb(iris)

a <- x$Species
b <- x$Petal_Length

c <- merge(a, b)
merge(b, b, merge=TRUE)


# Here is an example that joins on SciDB array attributes instead of
# dimensions using the 'equi_join' SciDB plugin operator.
set.seed(1)
a <- as.scidb(data.frame(a=sample(10, 5), b=rnorm(5)))
b <- as.scidb(data.frame(u=sample(10, 5), v=rnorm(5)))
merge(x=a, y=b, by.x="a", by.y="u")[]


# The following example joins on a subset of coordinate axes:
x <- build(5.5, c(3, 3));                   print(schema(x))
y <- build(1.1, c(3, 3), chunksize=c(2,1)); print(schema(y))
z <- merge(x, y, by="i")
print(schema(z))
}
}

