% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{names<-.scidb}
\alias{names<-.scidb}
\title{Renamed SciDB attributes and/or dimensions}
\usage{
\method{names}{scidb}(x) <- value
}
\arguments{
\item{x}{\code{scidb} array object}

\item{value}{a character vector with new dimension and attribute names, in that order}
}
\value{
\code{scidb} array object
}
\description{
Renamed SciDB attributes and/or dimensions
}

