% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{persist}
\alias{persist}
\title{Recursively set all arrays in a \code{scidb} dependency graph to persist.}
\usage{
persist(x, gc = FALSE)
}
\arguments{
\item{x}{a \code{\link{scidb}} object}

\item{gc}{\code{TRUE} connects arrays to R's garbage collector}
}
\description{
Recursively set all arrays in a \code{scidb} dependency graph to persist.
}

