% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.R
\name{predict.glm_scidb}
\alias{predict.glm_scidb}
\title{Predict method for SciDB GLM}
\usage{
\method{predict}{glm_scidb}(object, ...)
}
\arguments{
\item{object}{\code{glm_scidb} model object}

\item{...}{other prediction arguments including \code{newdata} (a \code{scidb} array of data to predict from),
\code{link} (a character value of either "link" or "response").}
}
\value{
\code{scidb} array object
}
\description{
Obtains predictions and optionally estimates standard errors of
those predictions from a fitted generalized linear model object.
}
\seealso{
\code{\link{predict.glm}}
}

