% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.R
\name{print.glm_scidb}
\alias{print.glm_scidb}
\title{Summary print method for SciDB GLM}
\usage{
print.glm_scidb(x, ...)
}
\arguments{
\item{...}{other arguments to \code{summary}}

\item{object}{\code{glm_scidb} model object}
}
\value{
Character summary of the model object is printed to standard output
}
\description{
Print a summary of a SciDB GLM object.
}
\seealso{
\code{\link{summary.glm}}
}

