% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{project}
\alias{project}
\title{Project onto a subset of SciDB array attributes}
\usage{
project(x, attributes)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{attributes}{a character vector of array attribute names or a numeric vector of attribute positions}
}
\value{
a new \code{scidb} array object
}
\description{
Project onto a subset of SciDB array attributes
}

