% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\docType{methods}
\name{qhyper,scidb-method}
\alias{qhyper,scidb-method}
\title{The Hypergeometric distribution}
\usage{
\S4method{qhyper}{scidb}(x, p, m, n, k, new = "q")
}
\arguments{
\item{x}{vector of quantiles representing the number of white balls
drawn without replacement from an urn which contains both
black and white balls.}

\item{p}{vector of quantiles representing the number of white balls
drawn without replacement from an urn which contains both
black and white balls.}

\item{m}{the number of white balls in the urn.}

\item{n}{the number of black balls in the urn.}

\item{k}{the number of balls drawn from the urn.}

\item{new}{the name of the new attribute with the resulting quantiles}
}
\value{
a new \code{scidb} array object
}
\description{
Density and quantile function for the hypergeometric distribution.
}
\seealso{
\code{\link{phyper}} \code{\link{qhyper}}
}

