% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{raw2scidb}
\alias{raw2scidb}
\title{Internal function to upload an R raw value to special 1-element SciDB array}
\usage{
raw2scidb(db, X, name, gc = TRUE, ...)
}
\arguments{
\item{db}{scidb database connection}

\item{X}{a raw value}

\item{name}{(character) SciDB array name}

\item{gc}{(boolean) set to \code{TRUE} to connect SciDB array to R's garbage collector}

\item{...}{optional extra arguments
\itemize{
\item {temp:} {(boolean) create a temporary SciDB array}
}}
}
\value{
a \code{\link{scidb}} object
}
\description{
Internal function to upload an R raw value to special 1-element SciDB array
}
\keyword{internal}
