% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{rbind.scidb}
\alias{rbind.scidb}
\title{Concatenate two SciDB arrays}
\usage{
rbind.scidb(x, y)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{y}{\code{scidb} array object}
}
\value{
\code{scidb} array object
}
\description{
Concatenate two SciDB arrays with common schema along their last dimension.
}

