% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{regrid,scidb-method}
\alias{regrid}
\alias{regrid,scidb-method}
\title{SciDB regrid decimation operator}
\usage{
\S4method{regrid}{scidb}(x, grid, expr)
}
\arguments{
\item{x}{SciDB array}

\item{grid}{a vector of grid sizes as long as \code{length(dimensions(x))}}

\item{expr}{optional aggregation function applied to every attribute on the grid, or a quoted SciDB aggregation expression}
}
\description{
SciDB regrid decimation operator
}
\note{
The default aggregation function used in case \code{expr} is left missing is \code{max}.
}
\examples{
\dontrun{
x <- as.scidb(iris)
regrid(x, 10, min)
}
}

