% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{remove_old_versions}
\alias{remove_old_versions}
\title{Remove old SciDB array versions}
\usage{
remove_old_versions(x)
}
\arguments{
\item{x}{\code{scidb} array object}
}
\value{
\code{NULL}
}
\description{
Remove old SciDB array versions
}

