% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{rename}
\alias{rename}
\title{Rename a SciDB array}
\usage{
rename(x, name = x@name, gc = FALSE)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{name}{optional new name}

\item{gc}{set \code{TRUE} to connect to result R's garbage collector (default is \code{FALSE})}
}
\value{
a \code{scidb} array object
}
\description{
Rename a SciDB array
}

