% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redim.R
\name{repart}
\alias{repart}
\title{SciDB repart operator}
\usage{
repart(x, schema, upper, chunk, overlap)
}
\arguments{
\item{x}{a \code{scidb} object}

\item{schema}{optional new schema}

\item{upper}{optional vector of new array coordinate dimensions}

\item{chunk}{optional vector of new array chunk sizes}

\item{overlap}{optional vector of new array chunk overlaps}
}
\value{
a \code{scidbf} object
}
\description{
SciDB repart operator
}

