% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{replaceNA}
\alias{replaceNA}
\title{SciDB 'substitute' operator, substitiute SciDB NULL values with new values}
\usage{
replaceNA(x, value, attribute)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{value}{value with which to replace \code{NULL} values}

\item{attribute}{Either a numeric vector of attribute positions or character vector of attribute names. Substitution is limited to the specified attributes.}
}
\value{
a new \code{scidb} array object
}
\description{
SciDB 'substitute' operator, substitiute SciDB NULL values with new values
}

