% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.R
\name{reshape_scidb}
\alias{reshape_scidb}
\title{SciDB reshape operator}
\usage{
reshape_scidb(x, schema, shape, dimnames, start, chunks)
}
\arguments{
\item{x}{a \code{scidb} object}

\item{schema}{optional new schema}

\item{shape}{optional vector of new array coordinate dimensions}

\item{dimnames}{optional new vector of array coordniate dimension name}

\item{start}{optional vector of new array starting coordinate index values}

\item{chunks}{optional vector of new array chunk sizes}
}
\value{
a \code{scidbf} object
}
\description{
SciDB reshape operator
}

