% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{schema}
\alias{schema}
\title{SciDB array schema}
\usage{
schema(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character-valued SciDB array schema
}
\description{
SciDB array schema
}

