% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidb}
\alias{scidb}
\title{Create an R reference to a SciDB array or expression.}
\usage{
scidb(name, gc = FALSE)
}
\arguments{
\item{name}{a character string name of a stored SciDB array or a valid SciDB AFL expression}

\item{gc}{a logical value, \code{TRUE} means connect the SciDB array to R's garbage collector}
}
\value{
a \code{scidb} object
}
\description{
Create an R reference to a SciDB array or expression.
}

