% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{scidb_arrow_to_dataframe}
\alias{scidb_arrow_to_dataframe}
\title{Return a SciDB query expression as a data frame}
\usage{
scidb_arrow_to_dataframe(db, query, ...)
}
\arguments{
\item{db}{scidb database connection object}

\item{query}{A SciDB query expression or scidb object}

\item{...}{optional extra arguments (see below)}
}
\description{
Return a SciDB query expression as a data frame
}
\note{
option extra arguments
\itemize{
  \item{only_attributes}{ optional logical value, TRUE if only attributes should be returned}
  \item{schema}{ optional result schema string }
}
}
\keyword{internal}
