% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_attributes}
\alias{scidb_attributes}
\title{SciDB array attribute names}
\usage{
scidb_attributes(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character vector of SciDB attribute names.
}
\description{
SciDB array attribute names
}

