% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_coordinate_bounds}
\alias{scidb_coordinate_bounds}
\title{SciDB array coordinate bounds}
\usage{
scidb_coordinate_bounds(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
list of character-valued vectors of starting and ending coordinate bounds
}
\description{
SciDB array coordinate bounds
}

