% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_coordinate_chunksize}
\alias{scidb_coordinate_chunksize}
\title{SciDB array coordinate chunksize}
\usage{
scidb_coordinate_chunksize(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character-valued vector of SciDB coordinate chunk sizes
}
\description{
SciDB array coordinate chunksize
}

