% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_coordinate_end}
\alias{scidb_coordinate_end}
\title{SciDB array coordinate end}
\usage{
scidb_coordinate_end(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character-valued vector of end coordinate bounds
}
\description{
SciDB array coordinate end
}

