% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_coordinate_overlap}
\alias{scidb_coordinate_overlap}
\title{SciDB array coordinate overlap}
\usage{
scidb_coordinate_overlap(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character-valued vector of SciDB coordinate overlap
}
\description{
SciDB array coordinate overlap
}

