% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_coordinate_start}
\alias{scidb_coordinate_start}
\title{SciDB array coordinate start}
\usage{
scidb_coordinate_start(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character-valued vector of starting coordinate bounds
}
\description{
SciDB array coordinate start
}

