% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{scidb_fisher.test}
\alias{scidb_fisher.test}
\title{Fisher's Exact Test for Count Data}
\usage{
scidb_fisher.test(a, x = "x", m = "m", n = "n", k = "k",
  alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{a}{SciDB array}

\item{x}{yes count attribute name}

\item{m}{marginal count attribute name}

\item{n}{marginal count attribute name}

\item{k}{amrginal count attribute name}

\item{alternative}{one of "two.sided", "greater", "less"}
}
\value{
A new SciDB array with two new attributes (note that the returned attribute
names may be adjusted to account for naming conflicts with existing array attributes):
pvalue (the p-value of the test), and estimate (an estimate of the odds ratio).
Note that the conditional maximum likelihood estimate (MLE) rather than the
unconditional MLE (the sample odds ratio) is used.
}
\description{
Performs Fisher's exact test for testing the null of independence
of rows and columns in a contingency table with fixed marginals.
For 2 by 2 tables, the null of conditional independence is equivalent to the
hypothesis that the odds ratio equals one. "Exact" inference can be based on
observing that in general, given all marginal totals fixed, the first element
of the contingency table has a non-central hypergeometric distribution with
non-centrality parameter given by the odds ratio (Fisher, 1935).
}
\details{
Consider the following 2x2 contingency table:
\tabular{lccc}{
\tab Class I YES\tab Class I NO\tab SUM\cr
Class II YES\tab x\tab a\tab k = x + a\cr
Class II NO\tab b\tab c\tab \cr
SUM\tab m = x + b\tab n = a + c\tab \cr
}
The \code{x} input value specifies the name of the SciDB array attribute that
indicates the number of 'yes' events in both classifications.
The \code{m} input value specifies the name of the SciDB array attribute that
indicates the marginal sum of the first column.
The \code{n} input value specifies the name of the SciDB array attribute that
indicates the marginal sum of the second column.
The \code{k} input value specifies the name of the SciDB array attribute that
indicates the marginal sum of the first row.
}
\examples{
\dontrun{
# Create a test array:
a <- scidb("apply(build(<x:int64>[i=0:0,1,0],2),m,12,n,18,k,17)")
scidb_fisher.test(a)[]

# output looks like:
#   x  m  n  k         pval   estimate
# 0 2 12 18 17 0.0005367241 0.04693664
}
}
\seealso{
\code{\link{scidb}} \code{\link{phyper}} \code{\link{qhyper}} \code{\link{dhyper}}
}

