% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_from_schemastring}
\alias{scidb_from_schemastring}
\title{SciDB array from a schema string}
\usage{
scidb_from_schemastring(s, expr = character())
}
\arguments{
\item{s}{schema character string}

\item{expr}{a character SciDB array name or array expression}
}
\value{
\code{\link{scidb}} reference object
}
\description{
SciDB array from a schema string
}
\note{
This function creates a virtual SciDB array object from a schema string and optional expression.
It does not contact the database nor does it create the array explicitly in SciDB. The function only
creates an R \code{\link{scidb}} array object with the indicated schema and expression.
}

