% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_nullable}
\alias{scidb_nullable}
\title{SciDB array attribute nullability}
\usage{
scidb_nullable(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
logical vector of SciDB attribute nullability
}
\description{
SciDB array attribute nullability
}

