% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema-utils.R
\name{scidb_types}
\alias{scidb_types}
\title{SciDB array attribute types}
\usage{
scidb_types(x)
}
\arguments{
\item{x}{a \code{\link{scidb}} array object}
}
\value{
character vector of SciDB attribute types
}
\description{
SciDB array attribute types
}

