% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbconnect}
\alias{scidbconnect}
\title{Connect to a SciDB database}
\usage{
scidbconnect(host = options("scidb.default_shim_host")[[1]],
  port = options("scidb.default_shim_port")[[1]], username, password,
  auth_type = "digest", protocol = c("http", "https"), init = FALSE)
}
\arguments{
\item{host}{optional host name or I.P. address of a SciDB shim service to connect to}

\item{port}{optional port number of a SciDB shim service to connect to}

\item{username}{optional authentication username}

\item{password}{optional authentication password}

\item{auth_type}{optional SciDB authentication type}

\item{protocol}{optional shim protocol type}

\item{init}{set to \code{TRUE} to instruct SciDB to load a lot of optional and required plug-ins once a connection is established}
}
\value{
\code{NULL} is returned invisibly; this function is used for its
side effect.
importFrom digest digest
}
\description{
Connect to a SciDB database
}
\note{
The SciDB connection state is maintained internally to the \code{scidb}
package. We internalize state to facilitate operations involving \code{scidb}
objects.

Thus, only one open SciDB connection is supported at
a time.

One may connect to and use multiple SciDB databases by sequentially calling
\code{scidbconnect} between operations. Note that \code{scidb} objects are not
valid across different SciDB databases.

Use the optional \code{username} and \code{password} arguments with
\code{auth_type} set to "digest" to use HTTP digest authentication (see the
shim documentation to configur this).  Digest authentication may use either
"http" or "https" transports selected by the \code{protocol} setting.

Disconnection is automatically handled by the package.
}

