% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidblist}
\alias{scidblist}
\title{List objects in a SciDB database}
\usage{
scidblist(pattern, type = c("arrays", "operators", "functions", "types",
  "aggregates", "instances", "queries", "libraries"), verbose = FALSE,
  n = Inf)
}
\arguments{
\item{pattern}{a regular-expression style search pattern}

\item{type}{(character), one of the indicated list types}

\item{verbose}{(boolean), include attribute and dimension data when type="arrays"}

\item{n}{maximum lines of output to return}
}
\value{
a list of matching SciDB database objects
}
\description{
List objects in a SciDB database
}

