% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbls}
\alias{scidbls}
\title{List objects in a SciDB database}
\usage{
scidbls(...)
}
\arguments{
\item{...}{options passed to \code{\link{scidblist}}}
}
\value{
a list
}
\description{
List objects in a SciDB database
}

