% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbrm}
\alias{scidbrm}
\title{Remove one or more scidb arrays}
\usage{
scidbrm(x, error = warning, ...)
}
\arguments{
\item{x}{a vector or single character string listing array names}

\item{error}{error handler. Use stop or warn, for example.}

\item{...}{additional options for \code{\link{scidbremove}}}
}
\value{
\code{NULL}
}
\description{
Remove one or more scidb arrays
}

