\name{show_commit_log}
\title{
Display the package change log
}
\description{
Display the fine-grained code commit log
for the currently installed version of the package.
}
\usage{
show_commit_log()
}
\value{
Prints the change log to standard output. Nothing is returned.
}
\author{
B. W. Lewis<blewis@paradigm4.com>
}
