% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{slice}
\alias{slice}
\title{SciDB 'slice' operator, slice out a particular dimension value}
\usage{
slice(x, d, n)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{d}{name or numeric position of dimension to slice}

\item{n}{coordinate value to slice on}
}
\value{
a new \code{scidb} array object
}
\description{
SciDB 'slice' operator, slice out a particular dimension value
}

