% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-functions.R
\name{[.scidb}
\alias{[.scidb}
\title{Projection onto array attributes (variables) or special return data to R operator}
\usage{
\method{[}{scidb}(x, ...)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{...}{character vector of array attributes}
}
\value{
\code{scidb} array object
}
\description{
Use empty indexing brackets to download the values of the SciDB array to R in a data frame.
}
\examples{
\dontrun{
x <- as.scidb(head(iris))
print(x)          # shows SciDB array structure
print(x[])        # downloads data back to R

x[, "Species"]    # data frame subset notation, returns a new SciDB array object
x[, "Species"][]  # Downloads to R
}
}
\seealso{
\code{\link{project}}
}

