% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{subarray}
\alias{subarray}
\title{SciDB 'subarray/between' operator, select subsets of arrays}
\usage{
subarray(x, limits, between = FALSE)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{limits}{vector of coordinate ranges to select}

\item{between}{use SciDB \code{between} operator instead of \code{subarray} when \code{TRUE}}
}
\value{
a new \code{scidb} array object
}
\description{
SciDB 'subarray/between' operator, select subsets of arrays
}

