% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.R
\name{summary.glm_scidb}
\alias{summary.glm_scidb}
\title{Summary method for SciDB GLM}
\usage{
summary.glm_scidb(object, ...)
}
\arguments{
\item{object}{\code{glm_scidb} model object}

\item{...}{other arguments to \code{summary}}
}
\value{
Character summary of the model object is printed to standard output and \code{NULL} is returned.
}
\description{
Print a summary of a SciDB GLM object.
}
\seealso{
\code{\link{summary.glm}}
}

