% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{t,scidb-method}
\alias{t,scidb-method}
\title{Matrix transpose}
\usage{
\S4method{t}{scidb}(x)
}
\arguments{
\item{x}{a 2-d scidb array with one numeric attribute (aka a matrix)}
}
\value{
a 2-d scidb array representing the transpose of \code{x}
}
\description{
Matrix transpose
}

