% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{transform.scidb}
\alias{transform.scidb}
\title{Transform SciDB array values}
\usage{
\method{transform}{scidb}(`_data`, ...)
}
\arguments{
\item{_data}{SciDB array}

\item{...}{named transformations, NOTE that SciDB expressions must be quoted (see example)}
}
\value{
a SciDB array
}
\description{
Use \code{transform} to add new derived attributes to a SciDB array, or to
replace an existing attribute. New attribute names must not conflict with array
dimension names.
}
\note{
The \code{transform} function at this time only supports quoted SciDB expressions.
A future version may also support more general mixed R/SciDB expressions similarly to
\code{\link{subset}}.
}
\examples{
\dontrun{
x <- scidb("build(<v:double>[i=1:5,5,0], i)")
y <- transform(x, a="2*v", v="3*v")
}
}

