% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{type_convert}
\alias{type_convert}
\title{Internal function that converts from SciDB types to R types}
\usage{
type_convert(x, types)
}
\arguments{
\item{x}{a data frame}

\item{types}{a string vector of SciDB type names}
}
\description{
Internal function that converts from SciDB types to R types
}
\keyword{internal}

