% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unbound}
\alias{unbound}
\title{Unbound SciDB array dimensions}
\usage{
unbound(x)
}
\arguments{
\item{x}{a \code{scidb} object}
}
\value{
a \code{scidb} object with unbounded dimension
}
\description{
This function sets each dimension upper bound to '*' (that is, unspecified)
}

