% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{unique.scidb}
\alias{unique.scidb}
\title{Filter unique elements from a SciDB array}
\usage{
\method{unique}{scidb}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{SciDB array}

\item{incomparables}{ignored}

\item{...}{optional arguments: specify \code{sort=TRUE} to first sort values}
}
\value{
a SciDB array
}
\description{
Filter unique elements from a SciDB array
}

