% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unordered_uniq}
\alias{unordered_uniq}
\title{Fast SciDB way to find unique elements}
\usage{
unordered_uniq(scidb_array, attributes)
}
\arguments{
\item{scidb_array}{a \code{scidb} array reference object, or a character string naming a stored SciDB array}

\item{attributes}{a list or character vector of SciDB array attributes}
}
\description{
A fast SciDB function for finding the unique elements among a set of attributes
in a SciDB array. This function uses the SciDB \code{\link{grouped_aggregate}} function.
}

