% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{unpack}
\alias{unpack}
\title{Unpack a SciDB array into a data frame-like shape}
\usage{
unpack(x)
}
\arguments{
\item{x}{\code{scidb} array object}
}
\value{
a new unpacked \code{scidb} array object
}
\description{
Unpack a SciDB array into a data frame-like shape
}

