% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{update.afl}
\alias{update.afl}
\title{Update available AFL operators}
\usage{
\method{update}{afl}(db, new, ops)
}
\arguments{
\item{db}{an \code{afl} object (a SciDB database connection returned from \code{\link{scidbconnect}})}

\item{new}{an optional list of new operators}

\item{ops}{an optional three-variable data frame with variables name, signature, help, corresponding
to the operator names, signatures, and help files (from SciDB Doxygen documentation)}
}
\value{
the updated database object
}
\description{
Update available AFL operators
}
\keyword{internal}

