% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{xgrid,scidb-method}
\alias{xgrid}
\alias{xgrid,scidb-method}
\title{SciDB xgrid prolongation operator}
\usage{
\S4method{xgrid}{scidb}(x, grid)
}
\arguments{
\item{x}{SciDB array}

\item{grid}{a vector of grid sizes as long as \code{length(dimensions(x))}}
}
\description{
SciDB xgrid prolongation operator
}
\examples{
\dontrun{
x <- as.scidb(iris)
y <- regrid(x, 10, min)
z <- xgrid(y, 10)
}
}

