% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload.R
\name{as.scidb}
\alias{as.scidb}
\title{Upload R data to SciDB
Data are uploaded as TSV.}
\usage{
as.scidb(X, name = tmpnam(), start, gc = TRUE, ...)
}
\arguments{
\item{X}{an R data frame, raw value, Matrix, matrix, or vector object}

\item{name}{a SciDB array name to use}

\item{start}{starting SciDB integer coordinate index}

\item{gc}{set to FALSE to disconnect the SciDB array from R's garbage collector}

\item{...}{other options, see \code{\link{df2scidb}}}
}
\value{
A \code{scidb} object
}
\description{
Upload R data to SciDB
Data are uploaded as TSV.
}

