% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{attribute_rename}
\alias{attribute_rename}
\title{Rename SciDB array attributes}
\usage{
attribute_rename(x, old, new)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{old}{Either a numeric vector of attribute positions, or a character vector of attribute names}

\item{new}{Character vector of new attribute names of same length as \code{old}}
}
\value{
a new \code{scidb} array object
}
\description{
Rename SciDB array attributes
}

