% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{build}
\alias{build}
\title{Wrapper to SciDB "build" operator}
\usage{
build(data, dim, names, type, start, name, chunksize, overlap, gc = TRUE)
}
\arguments{
\item{data}{any valid SciDB expression (expressed as a character string) or constant to fill the array}

\item{dim}{vector of dimension lengths}

\item{names}{optional vector of attribute and dimension names; default attribute name is \code{val}
and the dimension names are labeled \code{i, j, ...}}

\item{type}{SciDB type of the array attribute}

\item{start}{optional vector of starting dimension coordinate indices. Must match the length of the dim vector}

\item{name}{optional name of the SciDB array. An automatically generated name is used by default}

\item{chunksize}{optional vector of dimension chunk sizes. Must match the length of the dim vector}

\item{overlap}{optional vector of dimension overlap values. Must match the length of the dim vector}

\item{gc}{\code{TRUE} (the default) removes the array when corresponding R objects are garbage collected}
}
\value{
a \code{scidb} array object
}
\description{
The \code{build} function is a wrapper to the SciDB `build` operator.
Operation is similar to the R \code{matrix} and \code{array} functions.
It creates a new single-attribute SciDB array based on the specified parameters.
}
\examples{
\dontrun{
y <- build(pi, c(5, 3))
print(head(y))
}
}

