% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbops.R
\name{cast}
\alias{cast}
\title{SciDB cast operator}
\usage{
cast(x, schema)
}
\arguments{
\item{x}{\code{scidb} array object}

\item{schema}{new SciDB array schema to use}
}
\value{
a new \code{scidb} array object
}
\description{
SciDB cast operator
}

