% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{cumulate}
\alias{cumulate}
\title{SciDB Cumulative Aggregation
Use \code{cumulate} function to compute running operations along data,
for example cumulative sums.}
\usage{
cumulate(x, expression, dimension)
}
\arguments{
\item{x}{SciDB array}

\item{expression}{any valid SciDB aggregate expression, expressed as a character string}

\item{optional}{array dimension name (character) or 1-based dimension index to run along (default is to use the first dimension)}
}
\value{
SciDB array
}
\description{
SciDB Cumulative Aggregation
Use \code{cumulate} function to compute running operations along data,
for example cumulative sums.
}
\examples{
\dontrun{
x <- as.scidb(iris)
y <- cumulate(x, "sum(Petal_Width)")
}
}

