% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{GET_RAW}
\alias{GET_RAW}
\title{Basic HTTP GET request}
\usage{
GET_RAW(url)
}
\arguments{
\item{url}{a well-formed HTTP URL, will be url-encoded}
}
\value{
R raw binary HTTP content field
}
\description{
Basic HTTP GET request
}

