% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidb-methods.R
\docType{methods}
\name{hist,scidb-method}
\alias{hist,scidb-method}
\title{Histogram}
\usage{
\S4method{hist}{scidb}(x, breaks = 10, right = FALSE, materialize = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{scidb} object with a single numeric SciDB attribute}
}
\value{
an R histogram object
}
\description{
Return and optionall plot a histogram from SciDB array values
}
\seealso{
\code{\link{hist}}
}

