% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{iquery}
\alias{iquery}
\title{Run a SciDB query, optionally returning the result in a \code{data.frame}.}
\usage{
iquery(query, return = TRUE, n = -1)
}
\arguments{
\item{query}{a single SciDB query string}

\item{return}{if \code{TRUE}, return the result}

\item{n}{optional limit on number of results returned, set to -1 to return everything}
}
\description{
Run a SciDB query, optionally returning the result in a \code{data.frame}.
}

