% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{iquery}
\alias{iquery}
\title{Run a SciDB query, optionally returning the result.}
\usage{
iquery(db, query, return = FALSE, binary = TRUE, ...)
}
\arguments{
\item{db}{a scidb database connection from \code{\link{scidbconnect}}}

\item{query}{a single SciDB query string}

\item{return}{if \code{TRUE}, return the result}

\item{binary}{set to \code{FALSE} to read result from SciDB in text form}

\item{...}{additional options passed to \code{read.table} when \code{binary=FALSE}}
}
\description{
Run a SciDB query, optionally returning the result.
}

