% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redim.R
\name{redimension}
\alias{redimension}
\title{SciDB redimension operator wrapper function}
\usage{
redimension(x, schema, dim)
}
\arguments{
\item{x}{a \code{scidb} object}

\item{schema}{optional new schema}

\item{dim}{optional vector of dimension and attribute names to redimension along}
}
\value{
a \code{scidbf} object
}
\description{
SciDB redimension operator wrapper function
}
\note{
Redimension attributes must be of int64 type.
}

