% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.R
\name{redimension}
\alias{redimension}
\title{SciDB redimension operator wrapper function}
\usage{
redimension(x, schema, dim)
}
\arguments{
\item{x}{a \code{scidb} object}

\item{schema}{optional new schema}

\item{dim}{optional vector of dimension and/or attribute names to redimension along}
}
\value{
a \code{scidb} object
}
\description{
SciDB redimension operator wrapper function
}
\note{
Redimension attributes must be of int64 type.
}
\examples{
\dontrun{
x <- as.scidb(iris)
y <- redimension(transform(x, i="int64(Petal_Width * 10)"), dim="i")
z <- redimension(transform(x, i="int64(Petal_Width * 10)"), dim=c("row", "i"))
schema(x)
schema(y)
schema(z)
}
}

