% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.R
\name{redimension}
\alias{redimension}
\title{SciDB redimension operator}
\usage{
redimension(x, schema, dim)
}
\arguments{
\item{x}{a \code{scidbdf} object}

\item{schema}{optional new schema}

\item{dim}{optional vector of dimension and attribute names to redimension along}
}
\value{
a \code{scidbf} object
}
\description{
SciDB redimension operator
}

