% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jdbc.R
\name{scidbconnect}
\alias{scidbconnect}
\title{Connect to SciDB}
\usage{
scidbconnect(host = "127.0.0.1", port = 1239L, user, passwd,
  drv = system.file("java/scidb4j.jar", package = "scidb"))
}
\arguments{
\item{host}{I.P. address or host name}

\item{port}{SciDB server port number}

\item{user}{SciDB authentication}

\item{passwd}{SciDB authentication}

\item{drv}{full path to SciDB JDBC driver jar file}
}
\value{
JDBC database connection object is invisibly returned, connection is also stored in global package environment.
}
\description{
Connect to SciDB
}

