% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbconnect}
\alias{scidbconnect}
\title{Connect to a SciDB database}
\usage{
scidbconnect(host = getOption("scidb.default_shim_host", "127.0.0.1"),
  port = getOption("scidb.default_shim_port", 8080L), username, password,
  auth_type = c("scidb", "digest"), protocol = c("http", "https"))
}
\arguments{
\item{host}{optional host name or I.P. address of a SciDB shim service to connect to}

\item{port}{optional port number of a SciDB shim service to connect to}

\item{username}{optional authentication username}

\item{password}{optional authentication password}

\item{auth_type}{optional SciDB authentication type}

\item{protocol}{optional shim protocol type}
}
\value{
A scidb connection object.
}
\description{
Connect to a SciDB database
}
\note{
Use the optional \code{username} and \code{password} arguments with
\code{auth_type} set to "digest" to use HTTP digest authentication (see the
shim documentation to configure this).  Digest authentication may use either
"http" or "https" selected by the \code{protocol} setting.
Set \code{auth_type = "scidb"} to use SciDB authentication, which only
works over "https".

All arguments support partial matching.
}
\examples{
\dontrun{
s <- scidbconnect()

# SciDB 15.12 authentication example (using shim's default HTTPS port 8083)
s <- scidbconnect(user="root", password="Paradigm4", auth_type="scidb", port=8083)
}
}

