% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbdisconnect}
\alias{scidbdisconnect}
\title{Disconnect from a SciDB database}
\usage{
scidbdisconnect(db)
}
\description{
This explicitly releases the shim session assigned at `scidbconnect()`.
If not called, this is automatically called at garbage collection time
for a db object.
}
