% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbeval}
\alias{scidbeval}
\title{Evaluate an Expression to \code{scidb} or \code{scidb} objects}
\usage{
scidbeval(expr, eval = TRUE, name, gc = TRUE, temp = FALSE)
}
\arguments{
\item{name}{(character) optional SciDB array name to store as}

\item{gc}{(logical) optional, when TRUE tie result to R garbage collector}

\item{temp}{(logical, optional): when TRUE store as a SciDB temp array}

\item{exp}{A SciDB expression or \code{scidb} or \code{scidb} object}
}
\description{
Force evaluation of an expression that yields a \code{scidb} or \code{scidb} object,
storing the result to a SciDB array when \code{eval=TRUE}.
}

