% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{scidbremove}
\alias{scidbremove}
\title{Remove one or more scidb arrays}
\usage{
scidbremove(x, error = warning, force, warn = TRUE, recursive = FALSE)
}
\arguments{
\item{x}{(character), a vector or single character string listing array names}

\item{error}{(function), optional error handler function; use stop or warn, for example}

\item{force}{(optional boolean), if TRUE really remove this array, even if scidb.safe_remove=TRUE}

\item{warn}{(optional boolean), if FALSE supress warnings}

\item{recursive}{(optional boolean), if TRUE, recursively remove this array and its dependency graph}
}
\value{
\code{NULL}
}
\description{
Remove one or more scidb arrays
}

