% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbops.R
\name{transform.scidb}
\alias{transform.scidb}
\title{Transform a SciDB array using the SciDB 'apply' operator
Use \code{transform} to add new derived attributes to a SciDB array, or to
replace an existing attribute. New attribute names must not conflict with array
dimension names.}
\usage{
\method{transform}{scidb}(`_data`, ...)
}
\arguments{
\item{_data}{SciDB array}

\item{...}{named transformations, see example}
}
\value{
a SciDB array
}
\description{
Transform a SciDB array using the SciDB 'apply' operator
Use \code{transform} to add new derived attributes to a SciDB array, or to
replace an existing attribute. New attribute names must not conflict with array
dimension names.
}
\examples{
\dontrun{
x <- scidb("build(<v:double>[i=1:5,5,0], i)")
y <- transform(x, a=2*v, v=3*v)
}
}

